requirements.txt
vnstock
supabase
pandas
python-dotenv
requests
SQLAlchemy
psycopg2-binary
plotly
beautifulsoup4
matplotlib
seaborn
cloudscraper
tenacity
pdfplumber
pytesseract
pypdfium2
